Migrate to latest upstream Babel
Towards the end of 2017, the Babel maintainers simplified
how Babel distributed.[1] The recommended way to use it
now is with a new package whose name is expected to stay stable.
The package is now called
`babel-preset-env`

This PR is intended to silence runtime warnings that users now
encounter, as of Jan 2018, when they download and run Apollo
Launchpads with the 'Download' link.

More details at http://babeljs.io/env